+++ filter/f_glfw_window.h
@@ -96,8 +96,8 @@ inline void reformRtAsGl4x4(Mat & R, Mat & t, GLfloat * m)
 
 struct cmpglfwin{
   bool operator () (const GLFWwindow * a, const GLFWwindow * b){
-    return reinterpret_cast<unsigned int>(a) 
-      < reinterpret_cast<unsigned int>(b);
+    return reinterpret_cast<unsigned long long>(a) 
+      < reinterpret_cast<unsigned long long>(b);
   }
 };
